fixing

+++ backend/db/models/spot.js
@@ -38,10 +38,10 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat: {
-      type: DataTypes.DECIMAL
+      type: DataTypes.FLOAT
     },
     lng: {
-      type: DataTypes.DECIMAL
+      type: DataTypes.FLOAT
     },
     name: {
       type: DataTypes.STRING,
